--- src/executor/evaluate/functions/mod.rs
@@ -1,1 +1,0 @@
-
