+++ src/executor/evaluate/mod.rs
@@ -1,6 +1,5 @@
 mod error;
 mod evaluated;
-mod functions;
 
 use {
     super::{context::FilterContext, select::select},
